fix: harden flip repair invariants and PR quality gates
- Reject periodic external cells when replacement-cell parity constraints cannot be preserved.
- Canonicalize after D>=4 bulk repair failures before continuing with later insertions.
- Consolidate fixed-bug regression coverage into tests/regressions.rs and document the pattern.
- Narrow CodeQL/Codacy code scanning noise while keeping curated PR quality feedback via Codacy and CodeRabbit.
- Stage repository-owned Semgrep rules for future cleanup without enabling broad default rule packs.
