+++ .coderabbit.yml
@@ -106,9 +106,11 @@ reviews:
     yamllint:
       enabled: true
 
-    # Semantic code analysis
+    # Semantic code analysis is intentionally disabled until the staged,
+    # repository-owned rules in .semgrep.yaml are enabled after legacy cleanup.
+    # Default Semgrep packs are noisy here and duplicate CodeQL/cargo-audit.
     semgrep:
-      enabled: true
+      enabled: false
 
     # Python linter (ruff provides comprehensive Python analysis)
     ruff:
