+++ .github/workflows/codacy.yml
@@ -3,9 +3,8 @@
 # separate terms of service, privacy policy, and support
 # documentation.
 
-# This workflow checks out code, performs a Codacy security scan
-# and integrates the results with the
-# GitHub Advanced Security code scanning feature.
+# This workflow checks out code, runs Codacy's Markdownlint engine only, and
+# integrates the results with GitHub Advanced Security code scanning.
 # For more information on the Codacy security scan action usage and
 # parameters, see https://github.com/codacy/codacy-analysis-cli-action.
 # For more information on Codacy Analysis CLI in general, see
@@ -60,14 +59,12 @@ jobs:
           mkdir -p "$CODACY_WORKDIR"
           rsync -a --delete --exclude '.git' ./ "$CODACY_WORKDIR/"
 
-      # Execute Codacy Analysis CLI and generate a SARIF output with
-      # the security issues identified during the analysis
+      # Execute Codacy Analysis CLI with a single tool. The Codacy GitHub App may
+      # run curated PR-quality tools, but this SARIF workflow stays Markdownlint
+      # only so maintainability checks are not mirrored into GitHub Code Scanning.
       - name: Run Codacy Analysis CLI
         # Cap Codacy runtime so a hung analyzer does not consume the full job timeout.
         timeout-minutes: 20
-        # Codacy's opengrep/semgrep engine can fail intermittently;
-        # allow the workflow to continue so any generated SARIF is still uploaded.
-        continue-on-error: true
         uses: codacy/codacy-analysis-cli-action@562ee3e92b8e92df8b67e0a5ff8aa8e261919c08
         with:
           # Check https://github.com/codacy/codacy-analysis-cli#project-token
@@ -79,6 +76,7 @@ jobs:
           directory: ${{ env.CODACY_WORKDIR }}
           output: ${{ env.CODACY_SARIF }}
           format: sarif
+          tool: markdownlint
           skip-uncommitted-files-check: true
           # Adjust severity of non-security issues
           gh-code-scanning-compat: true
