+++ CONTRIBUTING.md
@@ -531,36 +531,57 @@ The project uses comprehensive CI workflows:
 
 - **Main CI** (`.github/workflows/ci.yml`): Build, test, lint on every PR
 - **Benchmarks** (`.github/workflows/benchmarks.yml`): Performance regression testing
-- **CodeQL** (`.github/workflows/codeql.yml`): GitHub code scanning for Actions, Python, and Rust
+- **CodeQL** (`.github/workflows/codeql.yml`): Security-focused GitHub code scanning for Actions, Python, and Rust
 - **Security** (`.github/workflows/audit.yml`): Dependency vulnerability scanning with SARIF upload
 - **Code Quality** (`.github/workflows/rust-clippy.yml`): Strict linting with SARIF upload
-- **Codacy** (`.github/workflows/codacy.yml`): Code quality analysis using project configurations and SARIF upload
+- **CodeRabbit** (`.coderabbit.yml`): PR review comments for curated quality feedback
+- **Codacy** (`.codacy.yml`): Curated PR quality feedback and duplication/complexity metrics
+- **Codacy SARIF mirror** (`.github/workflows/codacy.yml`): Markdownlint-only SARIF upload
 - **Coverage** (`.github/workflows/codecov.yml`): Test coverage tracking with 5-minute per-test timeout
 
 All PRs must pass CI checks before merging.
 
 ### 5. Code Quality Analysis
 
-The project uses **Codacy** for automated code quality analysis across both Rust and Python code:
+Non-security quality feedback should surface as PR review comments, normal
+status checks, or CI logs rather than broad GitHub Code Scanning alerts.
+
+The project uses **CodeRabbit** for PR review comments from the curated
+quality tool set:
+
+- **Rust**: Clippy
+- **Python**: Ruff
+- **Shell**: ShellCheck
+- **Markdown**: Markdownlint
+- **Configuration**: actionlint and yamllint
+- **Workflow context**: GitHub check summaries
+
+The project uses **Codacy** as a second PR-quality surface. Enable and disable
+Codacy tools in the Codacy Code Patterns UI; `.codacy.yml` records path and
+tool configuration, but Codacy does not use that file to turn tools on or off.
 
 - **Configuration**: `.codacy.yml` in the project root
-- **Rust Analysis**: Uses Clippy (configured via `Cargo.toml`) and rustfmt (configured via `rustfmt.toml`)
-- **Python Analysis**: Uses Ruff and Pylint (configured via `pyproject.toml`)
-- **Additional Tools**: ShellCheck for shell scripts, markdownlint for documentation, yamllint for config files
+- **Enabled Tools**: Markdownlint, Ruff, ShellCheck, duplication, and advisory Lizard
+- **Disabled Tools**: Bandit, Prospector, Pylint, broad Opengrep, Trivy, Jackson Linter, and Spectral
+- **Documentation Analysis**: Markdownlint uses `.markdownlint.json`
+- **Python Analysis**: Ruff uses `pyproject.toml`
+- **Local/CI Analysis**: Rust, Python, shell, YAML, TOML, JSON, and GitHub Actions checks run through `just check`
+- **Security Analysis**: Uses CodeQL and cargo-audit rather than Codacy's broader engine set
+- **Code Scanning Mirror**: `.github/workflows/codacy.yml` runs Markdownlint only so Codacy's maintainability
+  findings stay in PR feedback instead of GitHub Code Scanning
 
 **Key Benefits:**
 
-- **Unified Quality Dashboard**: Single view of code quality across all languages
-- **Uses Project Settings**: Respects your local tool configurations (no duplicate/conflicting rules)
-- **Pull Request Integration**: Quality feedback directly in PR reviews
-- **Code Scanning Integration**: SARIF results are uploaded to GitHub Code Scanning
-- **Trend Tracking**: Monitor code quality improvements over time
+- **Reduced Noise**: Avoids duplicate feedback from Bandit, Prospector, Pylint, broad Opengrep, Trivy,
+  Jackson Linter, and Spectral
+- **Uses Project Settings**: Respects repository Markdownlint and Ruff configuration
+- **Review Feedback**: Keeps maintainability comments close to the pull request
 
 **For Contributors:**
 
-- Codacy analysis runs automatically on all PRs
-- Quality issues are reported as PR comments
-- The same tools and rules used locally in development (following AGENTS.md guidelines)
+- CodeRabbit review feedback runs automatically on all PRs
+- Codacy analysis runs the curated quality tool set above
+- Broader lint and validation checks run locally and in CI via `just check`
 - No additional setup required - uses existing project configurations
 
 ## Commit Message Format
